refactor structure for queries
